Conflicts:
	include/ARX_Script.h
	include/EERIEApp.h
	include/EERIETypes.h
	include/HERMESMain.h
	include/hermes/PakEntry.h
	include/hermes/PakManager.h
	include/hermes/PakReader.h
	src/athena/Athena_FileIO.cpp
	src/athena/Athena_Stream_ASF.cpp
	src/danae/ARX_C_mapp.cpp
	src/danae/ARX_ChangeLevel.cpp
	src/danae/ARX_FTL.cpp
	src/danae/ARX_Interactive.cpp
	src/danae/ARX_Interface.cpp
	src/danae/ARX_Menu.cpp
	src/danae/ARX_Menu2.cpp
	src/danae/ARX_Menu2.h
	src/danae/ARX_Minimap.cpp
	src/danae/ARX_Player.cpp
	src/danae/ARX_Scene.cpp
	src/danae/ARX_Script.cpp
	src/danae/DanaeDlg.cpp
	src/danae/DanaeSaveLoad.cpp
	src/eerie/EERIEPoly.cpp
	src/eerie/EERIETexture.cpp
	src/eerie/EERIEobject.cpp
	src/hermes/HERMESmain.cpp
	src/hermes/PakEntry.cpp
	src/hermes/PakManager.cpp
	src/hermes/PakReader.cpp

+++ src/danae/ARX_ChangeLevel.cpp
@@ -63,7 +63,7 @@ ZeniMax Media Inc., Suite 120, Rockville, Maryland 20850 USA.
 //
 //////////////////////////////////////////////////////////////////////////////////////
 
-//-----------------------------------------------------------------------------
+#include <d3dwrapper.h>
 #include <vector>
 #include <algorithm>
 #include "ARX_ChangeLevel.h"
@@ -82,6 +82,7 @@ ZeniMax Media Inc., Suite 120, Rockville, Maryland 20850 USA.
 
 #include <HERMESMain.h>
 #include <hermes/PakManager.h>
+#include <hermes/Filesystem.h>
 
 #include <EERIEMath.h>
 #include <EERIEObject.h>
@@ -2116,7 +2117,7 @@ long ARX_CHANGELEVEL_Pop_Player(ARX_CHANGELEVEL_INDEX * asi, ARX_CHANGELEVEL_PLA
 	char _error[256];
 
 	sprintf(loadfile, "player.sav");
-	long size = _pSaveBlock->GetSize(loadfile);
+	size_t size = _pSaveBlock->GetSize(loadfile);
 
 	if (size <= 0)
 	{
@@ -2385,7 +2386,7 @@ long ARX_CHANGELEVEL_Pop_IO(char * ident)
 	ARX_CHANGELEVEL_IO_SAVE * ais;
 	unsigned char * dat;
 	long pos = 0;
-	long size = 0;
+	size_t size = 0;
 
 	sprintf(loadfile, "%s.sav", ident);
 
@@ -3471,7 +3472,7 @@ long ARX_CHANGELEVEL_Pop_Globals()
     unsigned char * dat;
     long pos = 0;
     std::string loadfile = "Globals.sav";
-    long size;
+    size_t size;
     std::string _error;
 
     ARX_SCRIPT_Free_All_Global_Variables();
@@ -4187,7 +4188,7 @@ long ARX_CHANGELEVEL_Get_Player_LevelData(ARX_CHANGELEVEL_PLAYER_LEVEL_DATA * pl
 
 	std::string loadfile;
 	std::string _error;
-	long size;
+	size_t size;
 	unsigned char * dat;
 
 	// Open Save Block
