+++ src/eerie/EERIEAnim.cpp
@@ -336,7 +336,7 @@ long EERIE_ANIMMANAGER_AddAltAnim(ANIM_HANDLE * ah,char * path)
 	if (!ah->path[0]) return 0;
 
 	unsigned char * adr;
-	long FileSize;
+	size_t FileSize;
 	EERIE_ANIM * temp;
 
 	if (!PAK_FileExist(path)) return 0;
@@ -372,7 +372,7 @@ ANIM_HANDLE * EERIE_ANIMMANAGER_Load( std::string& path)
 	}
 
 	unsigned char * adr;
-	long FileSize;
+	size_t FileSize;
 	char path2[256];
 	char pathcount=2;
